Estoy haciendo que se cree la carpeta del pan y que dentro del pan pueda crear las carpetas de los años

+++ Paginas/admi_general/mecanico/manuales/insertar_bd.php
@@ -12,16 +12,21 @@ if(isset($_POST['subir'])){
     $carpeta=htmlspecialchars($_POST['carpeta']); 
     $destino="pdf_manuales/";//carpeta de almacenamiento
     $directorio=$destino.$carpeta;//carpeta que contiene los manuales
-    mkdir($directorio, 0777);//crea el directorio
-    foreach($_FILES['manuales']['tmp_name'] as $key => $value){
-            $manuales=$_FILES['manuales']['tmp_name'];
-            if(file_exists($manuales[$key])){
-                move_uploaded_file($_FILES['manuales']['tmp_name'][$key],$directorio."/".$_FILES['manuales']['name'][$key]);
-                $ruta=$directorio."/".$_FILES['manuales']['name'][$key];
-            }
+    if(!file_exists($directorio)){
+        mkdir($directorio, 0777);//crea el directorio
+        foreach($_FILES['manuales']['tmp_name'] as $key => $value){
+                $manuales=$_FILES['manuales']['tmp_name'];
+                if(file_exists($manuales[$key])){
+                    move_uploaded_file($_FILES['manuales']['tmp_name'][$key],$directorio."/".$_FILES['manuales']['name'][$key]);
+                    $ruta=$directorio."/".$_FILES['manuales']['name'][$key];
+                }
+        }
+        $sql=$conexion->query("INSERT INTO manuales(carpeta) VALUES ('$carpeta')"); //guarda la ruta
+        header("Location: carga_equipo.php");
+    }else{
+        echo " $carpeta ya existe";
     }
-    $sql=$conexion->query("INSERT INTO manuales(carpeta) VALUES ('$carpeta')"); //guarda la ruta
-    header("Location: carga_equipo.php");
+    
 }
 
 ?>
